Include direcciones module, with services, screens and routes
