+++ src/app/modules/direccion/direccion-detalle/direccion-detalle.component.html
@@ -1,0 +1,48 @@
+<app-card-container [title]="'Direcciones'">
+    <div>
+        <div class="row align-items-center">
+            <div class="col-6">
+                <label># Dirección</label>
+                <input type="number" class="form-control" [(value)]="direccion.addressId" />
+            </div>
+            <div class="col-auto">
+                <label>Activo</label>
+                <input type="checkbox" [(checked)]="direccion.active" />
+            </div>
+        </div>
+
+        <div class="row">
+            <div class="col-6">
+                <label>Alias</label>
+                <input type="text" class="form-control" [(value)]="direccion.alias" />
+            </div>
+            <div class="col-6">
+                <label>Street</label>
+                <input type="text" class="form-control" [(value)]="direccion.street" />
+            </div>
+        </div>
+
+        <div class="row">
+            <div class="col-6">
+                <label>City</label>
+                <input type="text" class="form-control" [(value)]="direccion.city" />
+            </div>
+            <div class="col-6">
+                <label>State</label>
+                <input type="text" class="form-control" [(value)]="direccion.state" />
+            </div>
+        </div>
+
+        <div class="row">
+            <div class="col-6">
+                <label>ZipCode</label>
+                <input type="text" class="form-control" [(value)]="direccion.zipCode" />
+            </div>
+            <div class="col-6">
+                <label>country</label>
+                <input type="text" class="form-control" [(value)]="direccion.country" />
+            </div>
+        </div>
+
+    </div>
+</app-card-container>
